1 tests for sorting filtering

+++ sandbox.controller/src/kz/greetgo/sandbox/controller/model/ClientAccount.java
@@ -3,9 +3,9 @@ package kz.greetgo.sandbox.controller.model;
 import java.sql.Timestamp;
 
 public class ClientAccount {
-    public Integer id;
-    public Integer client;
-    public Float money;
+    public int id;
+    public int client;
+    public double money;
     public String number;
     public Timestamp registered_at;
 }
